fix: change providers to helper folder and add logger
